refactor: Update file paths for home components
