+++ app/(home)/during-the-stay-home.tsx
