+++ app/(home)/partners.tsx
@@ -1,6 +1,6 @@
 import dicoLogo from "@/assets/dico.png";
 import Image from "next/image";
-import { Button } from "../ui/button";
+import { Button } from "../../components/ui/button";
 import Link from "next/link";
 
 export function Partners() {
