+++ app/(home)/speakersScrollBar.module.css
